+++ gui/config.py
@@ -346,6 +346,7 @@ class ProcessingConfig:
     tbe_scene_cut_use_transnetv2: bool = False
     detection_denoise: bool = False
     sam2_refine: bool = False
+    matanyone_refine: bool = False
     edge_ring_px: int = 2
 
     subtitle_areas: Optional[List[Tuple[int, int, int, int]]] = None
@@ -498,6 +499,7 @@ class ProcessingConfig:
             self.tbe_scene_cut_use_transnetv2, False)
         self.detection_denoise = _coerce_bool(self.detection_denoise, False)
         self.sam2_refine = _coerce_bool(self.sam2_refine, False)
+        self.matanyone_refine = _coerce_bool(self.matanyone_refine, False)
         self.edge_ring_px = _coerce_int(self.edge_ring_px, 2, 0, 20)
         self.auto_band = _coerce_bool(self.auto_band, False)
         self.export_srt = _coerce_bool(self.export_srt, False)
@@ -887,6 +889,7 @@ SAFE_PRESET_FIELDS = frozenset({
     "tbe_scene_cut_use_transnetv2",
     "detection_denoise",
     "sam2_refine",
+    "matanyone_refine",
     "edge_ring_px",
     "auto_band",
     "adaptive_batch",